% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwhich.R
\name{\%forwhich\%}
\alias{\%forwhich\%}
\title{Return a single value from a vector, for which a different condition holds}
\usage{
x \%forwhich\% p
}
\arguments{
\item{x}{a vector from which to return results}

\item{p}{a vector of booleans (or other \code{[} indices?) length of x}
}
\description{
Intended to help with EAV calculations that potentially have length 0, ruining data frame operations.
}
\details{
Use case discussion here: https://community.rstudio.com/t/right-way-to-do-grouped-with-dplyr/1501
}
\examples{
tidyr::population \%>\%
  dplyr::group_by(country) \%>\%
  dplyr::summarize(chg_2000_to_2012 =
    (population \%forwhich\% year == 2012) /
    (population \%forwhich\% year == 2000),
    .groups = "drop")

}
