% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRAtools-package.R
\docType{package}
\name{JRAtools-package}
\alias{JRAtools}
\alias{JRAtools-package}
\title{JRAtools: Personal Utility Functions}
\description{
This provides utility functions that I tend to use in various projects.
}
\author{
\strong{Maintainer}: Jameel Alsalam \email{jalsalam@gmail.com} (\href{https://orcid.org/0000-0001-9801-5558}{ORCID})

}
\keyword{internal}
