% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwhich.R
\name{one}
\alias{one}
\title{Return the scalar of a vector}
\usage{
one(x)
}
\description{
Return the scalar of a vector
}
\examples{
one(c(1))
one(numeric(0))
typeof(one(numeric(0)))
\dontrun {
one(c(1, 2))
}
}
